fix: add component
