+++ src/pages/Create.tsx
@@ -1,50 +1,18 @@
 import { ImageUpload } from "../components/ImageUpload";
 import StoreInfo from "../components/StoreInfo";
 import * as S from "../styles/pages/Create.styles";
-import DeleteIcon from "../assets/cancel.svg";
 import { useState } from "react";
+import HashTag from "../components/HashTag";
 
 export function Create() {
   const [tagList, setTagList] = useState<string[]>([]);
-
-  const handleDelete = (tag: string) => {
-    setTagList(tagList.filter((t) => t !== tag));
-  };
-
-  const handleAddTag = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === "Enter") {
-      const newTag = event.currentTarget.value.trim();
-      if (newTag && !tagList.includes(newTag)) {
-        setTagList([...tagList, newTag]);
-        event.currentTarget.value = "";
-      }
-    }
-  };
 
   return (
     <S.CreateContainer>
       <S.LeftContainer>
         <StoreInfo />
         <ImageUpload />
-        <S.HashTagWrap>
-          <S.HashTagTitle>해시태그</S.HashTagTitle>
-          <S.HashTagInput
-            placeholder="해시태그를 입력하세요."
-            onKeyDown={handleAddTag}
-          />
-          <S.HashTagList>
-            {tagList.map((tag) => (
-              <S.HashTagTextBox key={tag}>
-                <S.HashTagText>#{tag}</S.HashTagText>
-                <S.HashTagDeleteButton
-                  src={DeleteIcon}
-                  alt="Delete"
-                  onClick={() => handleDelete(tag)}
-                />
-              </S.HashTagTextBox>
-            ))}
-          </S.HashTagList>
-        </S.HashTagWrap>
+        <HashTag tagList={tagList} setTagList={setTagList} />
       </S.LeftContainer>
       {/* right */}
     </S.CreateContainer>
